Move JSONAPIClient into tests.client module
